qrflow/requirements.txt
crispy-bootstrap5==0.7
Django==4.1.2
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-extensions==3.2.1
Faker==15.0.0
numpy==1.23.3
psycopg2==2.9.4
cryptography==38.0.1
qrcode==7.3.1
base45==0.4.4
cbor2==5.4.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
